+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## Unreleased
 
-- Removed the "Site" condition rule for non-localizable element types. ([#12601](https://github.com/craftcms/cms/pull/12601))
 - Fixed a bug where clicking on the scrollbar of a disclosure menu would close it. ([#12681](https://github.com/craftcms/cms/issues/12681))
 - Fixed a bug where large asset thumbnails weren’t centered within Assets fields. ([#12682](https://github.com/craftcms/cms/issues/12682))
+- Fixed a bug where the “Site” element condition rule was available for non-localizable element types’ conditions. ([#12601](https://github.com/craftcms/cms/pull/12601))
 
 ## 4.3.9 - 2023-02-14